rpm: Enable creation of RPMs
This is the first step.  I have confirmed that I can install,
update, and remove.  I have also setup the ability for Jenkins
to pass the BuildNumber and use that value as the release
number for the RPM.

The RPM depends on Grizzly(1.8.0) Swift from OpenStack.  To verify
you may need to add the appropiate repo file to your Fedora/RHEL system:
http://repos.fedorapeople.org/repos/openstack/openstack-grizzly

I have not had the opportunity to test that G4S itself works
once installed, but I plan on doing that as the next phase.

Change-Id: Ib90f335f5e1e4fc552c32e00ff29b6e8a680c42a
Signed-off-by: Luis Pabon <lpabon@redhat.com>
Reviewed-on: http://review.gluster.org/5006
Reviewed-by: Peter Portante <pportant@redhat.com>
Reviewed-by: Kaleb KEITHLEY <kkeithle@redhat.com>
Tested-by: Kaleb KEITHLEY <kkeithle@redhat.com>