Bring DiskFile module coverge to 100%
To bring the DiskFile module coverage to 100%, we first recognize that
do_fsync() will invoke fsync() in a separate thread, which gives
coverage fits (see the commit history at
https://github.com/portante/coverage/commits/master). To avoid that,
we mock out do_fsync to make it a no-op and avoid the problem.

The second thing we recognize is that mkstemp() relies on do_unlink
from the fs_utils module, which already consumes ENOENT errors, so we
don't need that code path in mkstemp() itself. The unused mock routine
for do_unlink was removed as well, and we renamed the other do_unlink
mock routine to os_unlink since it was mocking out os.unlink directly.

Lastly, we rejigger the error on close test for mkstemp() to
prematurely close the fd to cause an error which should just be
squelched, completing the full coverage.

Change-Id: I98283c17cf139f92282f8afd7083d567d3dd9a79
Signed-off-by: Peter Portante <peter.portante@redhat.com>
Reviewed-on: http://review.gluster.org/5082
Reviewed-by: Luis Pabon <lpabon@redhat.com>
Tested-by: Luis Pabon <lpabon@redhat.com>