Use tox to satisfy functional test requirements
Instead of installing and polluting the host development system
with the required packages to run the functional tests, it is
now possible to run the functional tests from inside tox.  All
the required packages are installed in the tox environment and
run from there.

The only external dependencies are that the functional tests
still must use /etc/swift for configuration files on the host
and /mnt/gluster-object must be on a XFS or GlusterFS volume.

Change-Id: I4a949dbcad778018e22c61d09ae1ba5f21a03aec
Signed-off-by: Luis Pabon <lpabon@redhat.com>
Reviewed-on: http://review.gluster.org/5976
Reviewed-by: Peter Portante <pportant@redhat.com>
Tested-by: Peter Portante <pportant@redhat.com>