Log only non-ENOENT errors in diskfile
DiskFile.open() gets called for every PUT.  The expected sideeffect
is that if the file exists, the metadata in the object will be
populated.

DiskFile.open() was logging any errors it detected while trying
to open() the file.  Issue is that when we PUT a new file, there
is no entry, and open() will return an error.  The was incorrectly
logging each one of these errors.

Change-Id: I7d721df177761066fdaa46d278fff2d779924999
Signed-off-by: Luis Pabon <lpabon@redhat.com>
Reviewed-on: http://review.gluster.org/6316
Reviewed-by: Thiago Da Silva <thiago@redhat.com>
Reviewed-by: Peter Portante <pportant@redhat.com>
Tested-by: Prashanth Pai <ppai@redhat.com>