Tests to verify locally deployed software.
The difference between these tests and those under the functional
directory, is that these tests do not run under tox, and so
require a full OpenStack Swift + swiftonfile deployment.

The conf files where copies of the gswauth test, but I modified them
to be able to do a mount check.

The purpose of this tests is to confirm the quality of the software
in a deployed system.

The tests require the following:
- GlusterFS Volumes: test, test2, gsmetadata
- Directory /mnt/gluster-object
- OpenStack Swift + swiftonfile to be installed.

Signed-off-by: Luis Pabon <lpabon@redhat.com>