clean up saio env after running func tests
Current tests from upstream swift leave containers in the
database after they are done running.

This causes an issue with sof tests because we clean up the
tests directory /mnt/swiftonfile. When running the
functional tests a second time always fails because the
setup functions tries to delete existing containers,
but the sof object servers returns 404, thus causing
the whole test to fail.

Signed-off-by: Thiago da Silva <thiago@redhat.com>