Fix problems on GET for files modified over FS
This patch attempts to correct problems where files modified from file
interface do not have metadata correctly updated, and consequently
problems occur with object download.

The problem at large is detecting whether a file was changed from
filesystem interface. This detection has to be done during precessing
of GET request.

This patch does the following:
* Introduce a new SOF specific metadata key named 'X-Object-PUT-Mtime'.
  During the last write() sys call, this key is populated with the
  file's current mtime (read: one inexpensive additional stat() call)
* Move object metadata creation and validation methods from utils.py
  to DiskFile class. The reasoning being:
  1) We can save stat and etag info as internal variables so that they
     can be reused, hence avoiding recomputation.
  2) The methods were only invoked from within DiskFile class and
     nowhere else. It makes sense to keep it close to where it's really
     being used.
* Override Swift's GET handler. If a file change is detected, issue a
  container_update() to update container database with newer timestamp,
  content length and etag.

For newer objects that do have 'X-Object-PUT-Mtime' key set in metadata,
we do not have to computer md5sum, all we do is compare current mtime
of file with the stored mtime in 'X-Object-PUT-Mtime' to check if file
has changed.

For existing objects in cluster which do not contain 'X-Object-PUT-Mtime'
key, we compute md5sum of file (read: expensive read()) and compare it
with etag stored as metadata to check if file has changed. This
recomputation will have to be done only once to populate the key
'X-Object-PUT-Mtime' properly.

Closes-Bug: #1416720
Co-Authored-By: Prashanth Pai <ppai@redhat.com>

Change-Id: I918f1bd5c9ba7839ea28e2d8ceec938c23f45413
Signed-off-by: Prashanth Pai <ppai@redhat.com>