Update validate_object to check size and etag
This patch attempts to correct problems where files modified from file interface
in swift on file container do not have metadata correctly updated, and consequently
problems occur with object download.

The updates modify validate_object, so that in addition to existing checks, the
mtime of the file is compared to the file metadata.  If it is are different,
the metadata is recalculated.  The assumption is that if the file changed it was
an intentional change in SoF and should be correctly reflected.

Closes-Bug: #1416720

Change-Id: I918f1bd5c9ba7839ea28e2d8ceec938c23f45413