Fix API incompatibility in update_metadata()
A major security fix[1] has been back-ported to icehouse branch in
upstream Swift. This fix changes the method signature of one of the
methods that we override, namely "update_metadata()". The method
signature of the same method in gluster-swift has been adjusted
accordingly. However, the entire fix itself has NOT been backported
as the implementations are different and due to time constraints.

[1]: https://review.openstack.org/#/c/126645/

Change-Id: I61d635ca5f47eb37f5275297c262f2bfe41de1fa
Signed-off-by: Prashanth Pai <ppai@redhat.com>