Exclude .trashcan dir from container listing
Since RHS 3.1, .trashcan directory is visible from glusterfs
fuse mount. This causes problems in gsexpiring volume in which
the directories should be integers (timestamps). The expirer
crashes when it tries to:

    int(".trashcan")

This change does not list .trashcan as a container when a GET
request is issued on the account/volume.

Change-Id: I70830a4085c5cf0890cb8e54ef8521c18a20c7f3
Signed-off-by: Prashanth Pai <ppai@redhat.com>