Optimize small file reads for new files
In case of newly added files or existing files modified over file interface,
the first GET on such files will result in the file being read twice - once
to compute Etag and again to send the file content to the client.

This is largely unavoidable as Etag is sent first as response header
followed by the response body which contains the actual object data.

However, small files can be cached the first time they're read and the
same be returned as response body later. This cached small file does not
persist across requests.

In this first implementation, the following is considered as a small file:
Any file whose entire content can be read in one read() sys call. This is
currently equal to 'disk_chunk_size' which by default is, 65536 bytes.

An unintended positive side effect of this is that, even in cases where
file is being appended or changed from file interface during the
processing of GET request, the Etag returned will be exact md5sum of
the response body. This wasn't the case earlier as the file may have
been modified between the first read and the second, resulting in
incorrect Etag.

Change-Id: Ic74941b10036baebc6ad3a897167684b1c5a5844
Signed-off-by: Prashanth Pai <ppai@redhat.com>