Don't (f)chown when it has no effect
For (f)chown calls which can change both UID and GID at once, -1 is
reserved as a sentinel value to indicate "omitted argument" or "do
not change". This makes sense when one of the args to (f)chown is -1.
When both uid and gid args are -1, it doesn't make sense to call
(f)chown as neither is going to be changed.

Further, as of today, diskfile doesn't get the information (uid and
gid) of the authenticated user from auth middleware. Retained the
calls in code for future when such functionality might be added.

This is a port of the following change:
http://review.gluster.org/#/c/13778

Change-Id: I74f1fa50bdfea30b404ac941b4345d8a94d5d141
Signed-off-by: Prashanth Pai <ppai@redhat.com>